backend/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
djoser==2.0.5
Markdown==3.3
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
django-filter